+++ src/sandbox/terminal.ts
